refactor: change files to camelCase pattern and repository approach to function instead of class
